+++ app/src/main/java/com/m2dl/cracotte/babal/scores/ScoreAdapter.java
@@ -1,4 +1,4 @@
-package com.m2dl.cracotte.babal.views;
+package com.m2dl.cracotte.babal.scores;
 
 import android.view.LayoutInflater;
 import android.view.View;
