+++ app/src/main/java/com/m2dl/cracotte/babal/scores/domain/ScoreTable.java
@@ -1,4 +1,6 @@
-package com.m2dl.cracotte.babal.entities;
+package com.m2dl.cracotte.babal.scores.domain;
+
+import com.m2dl.cracotte.babal.scores.domain.Score;
 
 import java.util.Map;
